requirements.txt
pyrogram
tgcrypto
python-dotenv
aiohttp
aiofiles
humanize
